feat: replace Apps Script with Drive comments
