+++ .env.example
@@ -1,5 +1,4 @@
 GROQ_API_KEY=your_groq_api_key_here
 GOOGLE_SERVICE_ACCOUNT_JSON=path/to/service_account.json
-GOOGLE_APPS_SCRIPT_ID=your_google_apps_script_id_here
 GROQ_CHUNK_SIZE=104857
 GROQ_REQUESTS_PER_MINUTE=10 # Groq allows about 10 RPM; higher values may be throttled
