+++ src/review.py
@@ -6,13 +6,13 @@ import logging
 from difflib import SequenceMatcher
 
 from .google_docs import chunk_paragraphs, get_document_paragraphs
-from .google_apps_script import create_comment
 from .google_drive import (
     download_revision_text,
     get_app_properties,
     get_share_message,
     update_app_properties,
     reply_to_comment,
+    create_comment,
 )
 
 
@@ -198,7 +198,6 @@ def review_document(
 
 def post_comments(
     drive_service: Any,
-    script_service: Any,
     document_id: str,
     items: List[Dict[str, str]],
 ) -> None:
@@ -234,7 +233,7 @@ def post_comments(
         parts = _chunk_content(content)
         # Post the first part anchored to the text range
         comment = create_comment(
-            script_service,
+            drive_service,
             document_id,
             parts[0],
             item.get("start_index"),
