+++ test/test_runner.py
@@ -7,7 +7,6 @@ from src.main import run_once
 def test_run_once_reviews_and_posts(monkeypatch):
     drive = MagicMock()
     docs = MagicMock()
-    script = MagicMock()
 
     monkeypatch.setattr(
         "src.main.list_recent_docs", lambda svc, since: [{"id": "1"}, {"id": "2"}]
@@ -25,14 +24,13 @@ def test_run_once_reviews_and_posts(monkeypatch):
 
     posted = []
 
-    def fake_post(drive_service, script_service, doc_id, items):
-        assert script_service is script
+    def fake_post(drive_service, doc_id, items):
         posted.append((doc_id, items))
 
     monkeypatch.setattr("src.main.post_comments", fake_post)
 
     since = datetime.utcnow()
-    new_since = run_once(drive, docs, script, since)
+    new_since = run_once(drive, docs, since)
 
     assert posted == [(
         "1",
